implemantacion de metodo restore para clientes, proveedores, tipos de madera, insumos y eventos

+++ resources/views/modulos/administrativo/proveedores/index.blade.php
@@ -4,12 +4,12 @@
 @section('submenu')
     @include('modulos.sidebars.costos-side')
 @endsection
-@section('content') 
-<div class="div container h-content ">        
-    <div class="row">            
+@section('content')
+<div class="div container h-content ">
+    <div class="row">
         <div class="col-12 col-sm-10 col-lg-6 mx-auto">
-            
-           
+
+
             <h1 class="display-6" >Proveedores</h1>
             <hr>
             <!-- Button trigger modal -->
@@ -22,7 +22,7 @@
                         - {{ $error }} <br>
                     @endforeach
                 </div>
-                
+
             @endif
             <!-- Modal Crea maquina-->
             <form action="{{ route('proveedores.store') }}" method="POST">
@@ -35,20 +35,20 @@
                         <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                         </div>
                         <div class="modal-body">
-                             
-                            <div class="card-body">                                                
-                                                   
+
+                            <div class="card-body">
+
                                 <div class="row mb-3">
                                     <label for="identificacion" class="col-md-4 col-form-label text-md-end">{{ __('Nit o cedula') }}</label>
                                     <div class="col-md-6">
-                                        <input  id="identificacion" 
-                                                type="text" 
-                                                class="form-control @error('identificacion') is-invalid @enderror" 
-                                                name="identificacion" value="{{ old('identificacion') }}" 
-                                                required 
-                                                autocomplete="identificacion" 
+                                        <input  id="identificacion"
+                                                type="text"
+                                                class="form-control @error('identificacion') is-invalid @enderror"
+                                                name="identificacion" value="{{ old('identificacion') }}"
+                                                required
+                                                autocomplete="identificacion"
                                                 autofocus>
-        
+
                                         @error('identificacion')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
@@ -57,20 +57,20 @@
                                     </div>
                                 </div>
 
-                                
+
                                 <div class="row mb-3">
                                     <label for="nombre" class="col-md-4 col-form-label text-md-end">{{ __('Representante legal') }}</label>
-        
+
                                     <div class="col-md-6">
-                                        <input  id="nombre" 
-                                                type="text" 
-                                                class="form-control @error('nombre') is-invalid @enderror text-uppercase" 
-                                                name="nombre" 
-                                                value="{{ old('name') }}" 
-                                                required 
-                                                autocomplete="nombre" 
+                                        <input  id="nombre"
+                                                type="text"
+                                                class="form-control @error('nombre') is-invalid @enderror text-uppercase"
+                                                name="nombre"
+                                                value="{{ old('name') }}"
+                                                required
+                                                autocomplete="nombre"
                                                 autofocus>
-        
+
                                         @error('nombre')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
@@ -78,18 +78,18 @@
                                         @enderror
                                     </div>
                                 </div>
-        
+
                                 <div class="row mb-3">
                                     <label for="razon_social" class="col-md-4 col-form-label text-md-end">{{ __('Razón social') }}</label>
                                     <div class="col-md-6">
-                                        <input  id="razon_social" 
-                                                type="text" 
-                                                class="form-control @error('razon_social') is-invalid @enderror text-uppercase" 
-                                                name="razon_social" value="{{ old('razon_social') }}" 
-                                                 
-                                                autocomplete="razon_social" 
+                                        <input  id="razon_social"
+                                                type="text"
+                                                class="form-control @error('razon_social') is-invalid @enderror text-uppercase"
+                                                name="razon_social" value="{{ old('razon_social') }}"
+
+                                                autocomplete="razon_social"
                                                 autofocus>
-        
+
                                         @error('razon_social')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
@@ -100,66 +100,66 @@
 
                                 <div class="row mb-3">
                                     <label for="direccion" class="col-md-4 col-form-label text-md-end">{{ __('Address') }}</label>
-        
+
                                     <div class="col-md-6">
-                                        <input  id="direccion" 
-                                                type="text" 
-                                                class="form-control @error('direccion') is-invalid @enderror text-uppercase" 
-                                                name="direccion" 
-                                                value="{{ old('direccion') }}" 
-                                                required 
+                                        <input  id="direccion"
+                                                type="text"
+                                                class="form-control @error('direccion') is-invalid @enderror text-uppercase"
+                                                name="direccion"
+                                                value="{{ old('direccion') }}"
+                                                required
                                                 autocomplete="direccion">
-        
+
                                         @error('direccion')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
                                             </span>
                                         @enderror
                                     </div>
-                                </div> 
+                                </div>
 
                                 <div class="row mb-3">
                                     <label for="telefono" class="col-md-4 col-form-label text-md-end">{{ __('Teléfono ') }}</label>
-        
+
                                     <div class="col-md-6">
-                                        <input id="telefono" 
-                                                type="number" 
-                                                class="form-control @error('telefono') is-invalid @enderror" 
-                                                name="telefono" 
-                                                value="{{ old('telefono') }}" 
-                                                required 
+                                        <input id="telefono"
+                                                type="number"
+                                                class="form-control @error('telefono') is-invalid @enderror"
+                                                name="telefono"
+                                                value="{{ old('telefono') }}"
+                                                required
                                                 autocomplete="telefono">
-        
+
                                         @error('telefono')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
                                             </span>
                                         @enderror
                                     </div>
-                                </div> 
-                                
+                                </div>
+
                                 <div class="row mb-3">
                                     <label for="email" class="col-md-4 col-form-label text-md-end">{{ __('Email') }}</label>
-        
+
                                     <div class="col-md-6">
-                                        <input  id="email" 
-                                                type="email" 
-                                                class="form-control @error('email') is-invalid @enderror " 
-                                                name="email" 
-                                                value="{{ old('email') }}" 
-                                                required 
+                                        <input  id="email"
+                                                type="email"
+                                                class="form-control @error('email') is-invalid @enderror "
+                                                name="email"
+                                                value="{{ old('email') }}"
+                                                required
                                                 autocomplete="email">
-        
+
                                         @error('email')
                                             <span class="invalid-feedback" role="alert">
                                                 <strong>{{ $message }}</strong>
                                             </span>
                                         @enderror
                                     </div>
-                                </div> 
+                                </div>
                         </div>
-                                
-                            
+
+
                         </div>
                         <div class="modal-footer">
                         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cerrar</button>
@@ -167,8 +167,8 @@
                         </div>
                     </div>
                     </div>
-                </div>   
-            </form>               
+                </div>
+            </form>
         </div>
         <!-- Tabla -->
 
@@ -176,12 +176,13 @@
             <thead>
                 <tr>
                     <th>ID</th>
-                    <th>Representante legal</th> 
+                    <th>Representante legal</th>
                     <th>Razon social</th>
                     <th>Direccion</th>
                     <th>Telefono</th>
                     <th>Email</th>
-                    <th>Calificacion</th>                   
+                    <th>Calificacion</th>
+                    <th>Estado</th>
                     <th>Acciones</th>
                 </tr>
             </thead>
@@ -195,23 +196,30 @@
                         <td>{{ $proveedor->direccion }}</td>
                         <td>{{ $proveedor->telefono }}</td>
                         <td>{{ $proveedor->email }}</td>
-                       
+
                         <td>{{ $proveedor->calificacion }}</td>
+                        <td>{!! $proveedor->deleted_at == '' ? '<span class="badge bg-primary">Activo</span>'  : '<span class="badge bg-secondary">Inactivo</span>' !!}</td>
                         <td>
                             <div class="d-flex align-items-center ">
-                                
-                                <button class="btn btn-sm btn-danger" onclick="eliminarUsuario({{ $proveedor }})">
-                                    <i class="fa-regular fa-trash-can fa-lg" style="color: black"></i>
-                                </button>
-                                <a href="{{ route('proveedores.show',$proveedor) }}" class="btn btn-sm btn-warning">
-                                    <i class="fa-solid fa-pen-to-square fa-lg"></i>
-                                </a>
-                               
+
+                                @if ($proveedor->deleted_at == '')
+                                    <button class="btn btn-sm btn-danger" onclick="eliminarUsuario({{ $proveedor }})">
+                                        <i class="fa-regular fa-trash-can fa-lg" style="color: black"></i>
+                                    </button>
+                                    <a href="{{ route('proveedores.show',$proveedor) }}" class="btn btn-sm btn-warning">
+                                        <i class="fa-solid fa-pen-to-square fa-lg"></i>
+                                    </a>
+                                @else
+                                    <button class="btn btn-sm btn-secondary" onclick="restaurarProveedor({{ $proveedor }})">
+                                        <i class="fa-solid fa-trash-can-arrow-up" style="color: black"></i>
+                                    </button>
+                                @endif
+
                             </div>
                         </td>
-                    </tr> 
+                    </tr>
                 @endforeach
-                
+
             </tbody>
         </table>
     </div>
